solved the issue of having to manually grant transaction

+++ my-flask-backend/blockchain/blockchain.py
@@ -1,7 +1,7 @@
 import json
 from web3 import Web3, HTTPProvider
 import json
-
+import os
 
 # truffle development blockchain address
 blockchain_address = 'http://localhost:9545'
@@ -49,30 +49,32 @@ print('*********************************')
 
 flag = 1
 while flag:
-    f=open('.\\formdata.json','r')
-    descr=f.readline()
-    prev_addr=descr
-    product_id=descr
-    # descr=input("description : ")
-    # prev_addr = input("enter the previous hash of blocks  used : ")
-    # product_id = input("enter the  product id : ")
-    trans = contract.functions.addBlock(descr,prev_addr,product_id).transact({'from':web3.eth.accounts[0]})
+    if os.path.exists('.\\formdata.json'):
+        f=open('.\\formdata.json','r')
+        descr=f.readline()
+        prev_addr=descr
+        product_id=descr
+        # descr=input("description : ")
+        # prev_addr = input("enter the previous hash of blocks  used : ")
+        # product_id = input("enter the  product id : ")
+        trans = contract.functions.addBlock(descr,prev_addr,product_id).transact({'from':web3.eth.accounts[0]})
 
-    web3.eth.wait_for_transaction_receipt(trans)
-    blockchain=contract.functions.getBlockchain().call()   
-    print('*********************************')
+        web3.eth.wait_for_transaction_receipt(trans)
+        blockchain=contract.functions.getBlockchain().call()   
+        print('*********************************')
 
-    print("updated blockchain : ")
+        print("updated blockchain : ")
 
-    data={}
-    i=1
-    for block in blockchain:
-        print(block)
-        data[i]=block
-        i+=1
+        data={}
+        i=1
+        for block in blockchain:
+            print(block)
+            data[i]=block
+            i+=1
 
-    with open('.\\bcdata.json', 'w') as f:
-        json.dump(data, f)
-    print('*********************************')
-
-    flag = int(input('proceed adding more ? if yes type 1 else type 0 : '))
+        with open('.\\bcdata.json', 'w') as f:
+            json.dump(data, f)
+        print('*********************************')
+        
+        os.remove('.\\formdata.json')
+    # flag = int(input('proceed adding more ? if yes type 1 else type 0 : '))
